unit test for actions/reducers [filter, selected-services]
