+++ api-catalog-ui/frontend/src/actions/selected-service-actions.test.jsx
@@ -1,25 +1,26 @@
 /* eslint-disable no-undef */
 
+import * as actions from './selected-service-actions'
 import { CLEAR_SERVICE, SELECT_SERVICE } from '../constants/selected-service-constants';
 
 describe('>>> Selected Service actions tests', () => {
     it('should return selected service', () => {
-        const state = { selectedService: {id: 'one'}, selectedTile: 'aaaa' };
-        const expectedState = { selectedService: {id: 'one'}, selectedTile: 'aaaa' };
-        const action = { type: 'SELECT_SERVICE', selectedService: {id: 'one'}, selectedTile: 'aaaa' };
-        expect(selectedServiceReducer(state, action)).toEqual(expectedState);
+        const selectedService = {id: 'service'};
+        const selectedTile = 'Tile';
+        const expectedAction = {
+            type: 'SELECT_SERVICE',
+            selectedService: {id: 'service'},
+            selectedTile: 'Tile'
+        };
+        expect(actions.selectService(selectedService, selectedTile)).toEqual(expectedAction);
     });
 
     it('should return clear service', () => {
-        const state = { selectedService: {id: 'one'}, selectedTile: 'aaaa' };
-        const expectedState = { selectedService: {}, selectedTile: null };
-        const action = { type: 'CLEAR_SERVICE'};
-        expect(selectedServiceReducer(state, action)).toEqual(expectedState);
-    });
-
-    it('should return default state', () => {
-        const state = { selectedService: {id: 'one'}, selectedTile: 'aaaa' };
-        const action = { type: 'OOPS'};
-        expect(selectedServiceReducer(state, action)).toEqual(state);
+        const expectedAction = {
+            type: CLEAR_SERVICE,
+            selectedService: {},
+            selectedTile: ""
+        };
+        expect(actions.clearService()).toEqual(expectedAction);
     });
 });
