+++ api-catalog-ui/frontend/src/epics/fetch-tiles.jsx
@@ -4,7 +4,7 @@ import { ofType } from 'redux-observable';
 import { catchError, debounceTime, exhaustMap, map, mergeMap, retryWhen, takeUntil } from 'rxjs/operators';
 import { FETCH_TILES_REQUEST, FETCH_TILES_STOP } from '../constants/catalog-tile-constants';
 import { fetchTilesFailed, fetchTilesRetry, fetchTilesSuccess } from '../actions/catalog-tile-actions';
-import { userActions } from '../actions/user.actions';
+import { userActions } from '../actions/user-actions';
 
 const updatePeriod = Number(process.env.REACT_APP_STATUS_UPDATE_PERIOD);
 const debounce = Number(process.env.REACT_APP_STATUS_UPDATE_DEBOUNCE);
