+++ api-catalog-ui/frontend/src/reducers/filter-reducer.test.jsx
@@ -4,24 +4,6 @@ import { CLEAR_FILTER, FILTER_TEXT } from '../constants/filter-constants';
 import filtersReducer from './filter-reducer';
 
 describe('>>> Filter reducer tests', () => {
-    it('should create actions to clear text', () => {
-        const expectedAction = {
-            type: CLEAR_FILTER,
-            defaultFilter: '',
-        };
-
-        expect(clear()).toEqual(expectedAction);
-    });
-
-    it('should create actions to filter text', () => {
-        const expectedAction = {
-            type: FILTER_TEXT,
-            text: 'test',
-        };
-
-        expect(filterText('test')).toEqual(expectedAction);
-    });
-
     it('should handle CLEAR_FILTER', () => {
         const expectedState = {
             text: '',
